fix(web): anchor auth callback URLs to web origin for Core client
Relative callbackURL values resolve against the Core auth server when
NEXT_PUBLIC_USE_CORE_AUTH_CLIENT is enabled, sending users to api.preprod
after magic link, email sign-in, and similar flows. Add
getAbsoluteAuthRedirectUrl and consolidate redirect helpers in auth.utils.
