+++ apps/web/src/app/(app)/connections/components/social-accounts.tsx
@@ -9,6 +9,7 @@ import { GoogleIcon, MicrosoftIcon } from "@/components/social-icons";
 import { Button } from "@/components/ui/button";
 import type { Account } from "@/lib/auth/auth";
 import { authClient } from "@/lib/auth/auth.client";
+import { getAbsoluteAuthRedirectUrl } from "@/lib/auth/auth.utils";
 import { AccountProvider } from "@/lib/auth/types";
 
 import DisconnectModal from "./disconnect-modal";
@@ -38,7 +39,7 @@ export function SocialAccounts({ socialAccounts }: SocialAccountsProps) {
     setLoading(true);
     const result = await authClient.linkSocial({
       provider,
-      callbackURL: "/connections",
+      callbackURL: getAbsoluteAuthRedirectUrl("/connections"),
     });
     if (result.error) {
       const errorMessage = result.error.message ?? t("error", { provider });
