+++ apps/web/src/app/(auth)/signin/components/form.tsx
@@ -21,7 +21,7 @@ import {
   buildSignUpUrlFromSignIn,
   createAuthSessionGetter,
   getAuthOAuthRedirect,
-  getValidAuthRedirectUrl,
+  getAbsoluteAuthRedirectUrl,
   normalizeAuthReturnUrl,
   waitForAuthSession,
 } from "@/lib/auth/auth.utils";
@@ -81,7 +81,7 @@ export default function SignInForm({
       email: values.email,
       password: values.currentPassword,
       rememberMe: values.rememberMe,
-      callbackURL: getValidAuthRedirectUrl(effectiveReturnUrl, "/"),
+      callbackURL: getAbsoluteAuthRedirectUrl(effectiveReturnUrl, "/"),
     });
 
     if (result.error) {
