+++ apps/web/src/lib/utils/url.ts
@@ -1,6 +1,4 @@
 import { siteConfig } from "@/config/site";
-import { getValidAuthRedirectUrl } from "@/lib/auth/auth.utils";
-import type { SocialProviderId } from "@/lib/schemas";
 
 export function getHostname(rawUrl: string): string | null {
   try {
@@ -45,36 +43,6 @@ export function getReturnUrlFromCurrentLocation(): string {
       ? window.location.pathname + window.location.search
       : "/chat";
   return path === "/" || path === "" ? "/chat" : path;
-}
-
-/**
- * Builds a social-auth callback URL for `authClient.signIn.social`.
- *
- * The result is an **absolute** URL anchored to the current web origin. This
- * matters when the browser `authClient` targets the Core Better Auth instance
- * (a different origin, e.g. `api.preprod.sokosumi.com`): Better Auth resolves a
- * relative `callbackURL` against the auth-server origin, so a bare
- * `/auth/callback/signin` would both land on the Core domain and collide with
- * Core's own `/auth/callback/:provider` route — surfacing as `state_not_found`.
- * Anchoring to `window.location.origin` (already a trusted origin) sends the
- * user back to the web app after the OAuth callback completes. Falls back to a
- * relative path when `window` is unavailable (SSR).
- */
-export function buildAuthCallbackUrl(
-  path: string,
-  provider: SocialProviderId,
-  returnUrl?: string,
-): string {
-  const params = new URLSearchParams({ provider });
-  if (returnUrl) {
-    const safeReturnUrl =
-      typeof window !== "undefined"
-        ? getValidAuthRedirectUrl(returnUrl, "/")
-        : returnUrl;
-    params.set("returnUrl", safeReturnUrl);
-  }
-  const origin = typeof window !== "undefined" ? window.location.origin : "";
-  return `${origin}${path}?${params.toString()}`;
 }
 
 export function buildJobTransactionUrl(
